add outline text effect

+++ core/font/font.h
@@ -30,7 +30,7 @@ typedef struct ex_font_t {
     void *face;
     int size;
     int outline_type; // ( 0 = None, 1 = Line, 2 = Inner, 3 = Outer )
-    int outline_thickness;
+    float outline_thickness;
     ex_hashmap_t *glyph_sets; // (size, outline_type, outline_thickness) -> ex_glyph_set_t 
 } ex_font_t;
